refactor package name
add getter setter for barangay name, town name, province name
