+++ src/main/java/org/guanzon/cas/model/clients/Model_Client_Mail.java
@@ -1,35 +1,42 @@
-package org.guanzon.clients;
+package org.guanzon.cas.model.clients;
 
 import java.lang.reflect.Method;
 import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.sql.Statement;
 import java.util.Date;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import javax.sql.rowset.CachedRowSet;
+import javax.sql.rowset.RowSetProvider;
 import org.guanzon.appdriver.base.GRider;
 import org.guanzon.appdriver.base.MiscUtil;
 import org.guanzon.appdriver.base.SQLUtil;
 import org.guanzon.appdriver.constant.EditMode;
 import org.guanzon.appdriver.constant.Logical;
+import org.guanzon.appdriver.constant.RecordStatus;
 import org.guanzon.appdriver.iface.GEntity;
 import org.json.simple.JSONObject;
+
 
 /**
  *
  * @author Michael Cuison
  */
-public class Model_Client_Social_Media implements GEntity{
+public class Model_Client_Mail implements GEntity{
     
-    final String XML = "Model_Client_Social_Media.xml";
+    final String XML = "Model_Client_Mail.xml";
     Connection poConn;          //connection
     CachedRowSet poEntity;      //rowset
     String psMessage;           //warning, success or error message
+    
+    
     GRider poGRider;
     int pnEditMode;
     public JSONObject poJSON;
-    public Model_Client_Social_Media(Connection foValue, GRider poValue){
+    
+    public Model_Client_Mail(Connection foValue, GRider poValue){
         if (foValue == null){
             System.err.println("Database connection is not set.");
             System.exit(1);
@@ -37,10 +44,12 @@ public class Model_Client_Social_Media implements GEntity{
         pnEditMode = EditMode.UNKNOWN;
         poGRider = poValue;
         poConn = foValue;
+        
         initialize();
     }
 
-    @Override
+    
+        @Override
     public String getColumn(int fnCol) {
         try {
             return poEntity.getMetaData().getColumnLabel(fnCol); 
@@ -61,41 +70,17 @@ public class Model_Client_Social_Media implements GEntity{
 
     @Override
     public int getColumnCount() {
-        try {
-            return poEntity.getMetaData().getColumnCount(); 
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-        
-        return -1;
+        throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
     }
 
+    /** 
+     * this is your table name
+     **/
     @Override
     public String getTable() {
-        return "Client_Social_Media";
+        return "Client_eMail_Address";
     }
-
-    @Override
-    public Object getValue(int fnColumn) {
-        try {
-            return poEntity.getObject(fnColumn);
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-        return null;
-    }
-
-    @Override
-    public Object getValue(String fsColumn) {
-        try {
-            return poEntity.getObject(fsColumn);
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-        return null;
-    }
-    
-    @Override
+     @Override
     public void list() {
         Method[] methods = this.getClass().getMethods();
         
@@ -105,20 +90,20 @@ public class Model_Client_Social_Media implements GEntity{
         }
     }
     /**
-     * Sets the ID of this record.
+     * Sets the Email ID of this record.
      * 
      * @param fsValue 
      * @return  True if the record assignment is successful.
      */
-    public JSONObject setSocialID(String fsValue){
-        return setValue("sSocialID", fsValue);
+    public JSONObject setEmailID(String fsValue){
+        return setValue("sEmailIDx", fsValue);
     }
     
-    /** 
-     * @return The ID of this record. 
+    /**
+     * @return The Email ID of this record. 
      */
-    public String getSocialID(){
-        return (String) getValue("sSocialID");
+    public String getEmailID(){
+        return (String) getValue("sEmailIDx");
     }
     
     /**
@@ -139,75 +124,56 @@ public class Model_Client_Social_Media implements GEntity{
     }
     
     /**
-     * Sets the social account.
+     * Sets the Entry no.
      * 
      * @param fsValue 
      * @return  True if the record assignment is successful.
      */
-    public JSONObject setSocialAccount(String fsValue){
-        return setValue("sAccountx", fsValue);
+    public JSONObject setEntryNo(int fsValue){
+        return setValue("nEntryNox", fsValue);
     }
     
     /**
-     * @return The social account.
+     * @return The Entry no.
      */
-    public String getSocialAccount(){
-        return (String) getValue("sAccountx");
+    public int getEntryNo(){
+        return (int) getValue("nEntryNox");
     }
     
-    /**
-     * Sets the remarks.<br>
+     /**
+     * Sets the email address.
      * 
      * @param fsValue 
      * @return  True if the record assignment is successful.
      */
-    public JSONObject setRemarks(String fsValue){
-        return setValue("sRemarksx", fsValue);
+    public JSONObject setEmail(String fsValue){
+        return setValue("sEMailAdd", fsValue);
     }
     
     /**
-     * @return The remarks. 
+     * @return The email address.
      */
-    public String getRemarks(){
-        return (String) getValue("sRemarksx");
+    public String getEmail(){
+        return (String) getValue("sEMailAdd");
     }
     
     /**
-     * Sets the social account type.<br>
-     * 0 for Facebook<br>
-     * 1 for Instagram<br>
-     * 2 for Twitter
+     * Sets the priority.
      * 
      * @param fsValue 
      * @return  True if the record assignment is successful.
      */
-    public JSONObject setSocialTyp(String fsValue){
-        return setValue("cSocialTp", fsValue);
+    public JSONObject setPriority(int fsValue){
+        return setValue("nPriority", fsValue);
     }
     
     /**
-     * @return The social account type. 
+     * @return The priority.
      */
-    public String getSocialTyp(){
-        return (String) getValue("cSocialTp");
+    public int getPriority(){
+        return (int) getValue("nPriority");
     }
     
-    /**
-     * Sets record as active.
-     * 
-     * @param fbValue 
-     * @return  True if the record assignment is successful.
-     */
-    public JSONObject setActive(boolean fbValue){
-        return setValue("cRecdStat", fbValue ? "1" : "0");
-    }
-    
-    /**
-     * @return If record is active. 
-     */
-    public boolean isActive(){
-        return ((String) getValue("cRecdStat")).equals("1");
-    }
     
     /**
      * Sets the user encoded/updated the record.
@@ -243,38 +209,49 @@ public class Model_Client_Social_Media implements GEntity{
         return (Date) getValue("dModified");
     }
     
-    public String getMessage(){
-        return psMessage;
+    @Override
+    public Object getValue(int fnColumn) {
+        try {
+            return poEntity.getObject(fnColumn);
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return null;
     }
-    
+
+    @Override
+    public Object getValue(String fsColumn) {
+        try {
+            return poEntity.getObject(fsColumn);
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return null;
+    }
+
     private String getSQL(){
         return "SELECT" +
-                    "  sSocialID" +
                     ", sClientID" +
-                    ", sAccountx" +
-                    ", sRemarksx" +
-                    ", cSocialTp" +
-                    ", cRecdStat" +
-                    ", sModified" +
-                    ", dModified" +
+                    ", nEntryNox" +
+                    ", sEmailAdd" +
+                    ", nPriority" +
                 " FROM " + getTable();
     }
-    
     private void initialize(){
         
         try {
             poEntity = MiscUtil.xml2ResultSet(System.getProperty("sys.default.path.metadata") + XML, getTable());
             
-            
             poEntity.last();
             poEntity.moveToInsertRow();
 
-            MiscUtil.initRowSet(poEntity);
-         //replace with the primary key column info
-            poEntity.updateString("sSocialID", MiscUtil.getNextCode(getTable(), "sSocialID", true, poConn, poGRider.getBranchCode()));
-            poEntity.updateString("cSocialTp", Logical.NO);
-            poEntity.updateString("cRecdStat", Logical.YES);
+            MiscUtil.initRowSet(poEntity);  
             
+            
+            poEntity.updateString("sEmailIDx", MiscUtil.getNextCode(getTable(), "sEmailIDx", true, poConn, poGRider.getBranchCode()));
+            poEntity.updateString("cPrimaryx", Logical.NO);
+            poEntity.updateString("cOwnerxxx", Logical.NO);
+            poEntity.updateString("cRecdStat", Logical.YES);
             poEntity.insertRow();
             poEntity.moveToCurrentRow();
 
@@ -283,15 +260,15 @@ public class Model_Client_Social_Media implements GEntity{
             e.printStackTrace();
             System.exit(1);
         }
-    }    
+    } 
     
-
+  
     @Override
     public JSONObject newRecord() {
         pnEditMode = EditMode.ADDNEW;
         
         //replace with the primary key column info
-        setSocialID(MiscUtil.getNextCode(getTable(), "sSocialID", true, poGRider.getConnection(), poGRider.getBranchCode()));
+        setEmailID(MiscUtil.getNextCode(getTable(), "sEmailIDx", true, poGRider.getConnection(), poGRider.getBranchCode()));
         
         poJSON = new JSONObject();
         poJSON.put("result", "success");
@@ -299,11 +276,11 @@ public class Model_Client_Social_Media implements GEntity{
     }
 
     @Override
-    public JSONObject openRecord(String fsValue) {
+    public JSONObject openRecord(String fsValue) { pnEditMode = EditMode.UPDATE;
         poJSON = new JSONObject();
 
         String lsSQL = MiscUtil.makeSelect(this);
-        lsSQL = MiscUtil.addCondition(lsSQL, "sSocialID = " + SQLUtil.toSQL(fsValue));
+        lsSQL = MiscUtil.addCondition(lsSQL, "sEmailIDx = " + SQLUtil.toSQL(fsValue));
 
         ResultSet loRS = poGRider.executeQuery(lsSQL);
 
@@ -337,7 +314,8 @@ public class Model_Client_Social_Media implements GEntity{
             String lsSQL;
             if (pnEditMode == EditMode.ADDNEW){
                 //replace with the primary key column info
-                setSocialID(MiscUtil.getNextCode(getTable(), "sSocialID", true, poGRider.getConnection(), poGRider.getBranchCode()));
+                setEmailID(MiscUtil.getNextCode(getTable(), "sEmailIDx", true, poGRider.getConnection(), poGRider.getBranchCode()));
+                
                 setModifiedDate(poGRider.getServerDate());
                 lsSQL = MiscUtil.makeSQL(this);
                 
@@ -354,15 +332,15 @@ public class Model_Client_Social_Media implements GEntity{
                     poJSON.put("message", "No record to save.");
                 }
             } else {
-                Model_Client_Social_Media loOldEntity = new Model_Client_Social_Media(poConn, poGRider);
+                Model_Client_Mail loOldEntity = new Model_Client_Mail(poConn, poGRider);
                 
-                //replace with the primary key column info
-                JSONObject loJSON = loOldEntity.openRecord(this.getSocialID());
                 setModifiedDate(poGRider.getServerDate());
+                //replace with the primary key column info
+                JSONObject loJSON = loOldEntity.openRecord(this.getEmailID());
                 
                 if ("success".equals((String) loJSON.get("result"))){
                     //replace the condition based on the primary key column of the record
-                    lsSQL = MiscUtil.makeSQL(this, loOldEntity, "sSocialID = " + SQLUtil.toSQL(this.getClientID()));
+                    lsSQL = MiscUtil.makeSQL(this, loOldEntity, "sEmailIDx = " + SQLUtil.toSQL(this.getEmailID()));
                     
                     if (!lsSQL.isEmpty()){
                         if (poGRider.executeQuery(lsSQL, getTable(), poGRider.getBranchCode(), "") > 0){
@@ -386,10 +364,48 @@ public class Model_Client_Social_Media implements GEntity{
             poJSON.put("message", "Invalid update mode. Unable to save record.");
             return poJSON;
         }
+        
         return poJSON;
-
+//
+//        poJSON =  new JSONObject();
+//        try {
+//            lsSQL = MiscUtil.rowset2SQL(poEntity, 
+//                    getTable(),
+//                    "",
+//                    "");
+//        
+//            if (pnEditMode == EditMode.ADDNEW){           
+//                lsSQL = MiscUtil.getNextCode(getTable(), "sEmailIDx", false, poGRider.getConnection(), "");
+//                poEntity.updateObject("sEmailIDx", lsSQL);
+//                poEntity.updateRow();
+//
+//                lsSQL = MiscUtil.rowset2SQL(poEntity, getTable(), "");
+//            } else {            
+//                lsSQL = MiscUtil.rowset2SQL(poEntity, 
+//                                            getTable(), 
+//                                            "", 
+//                                            "sEmailIDx = " + SQLUtil.toSQL(poEntity.getString("sEmailIDx")));
+//            }
+//            
+//            if (!lsSQL.equals("")){
+//                if(poGRider.executeQuery(lsSQL, getTable(), "", "") == 0){
+//                    if(!poGRider.getErrMsg().isEmpty()){ 
+//                        poJSON.put("result", "error");
+//                        poJSON.put("message", poGRider.getErrMsg());
+//                        return poJSON;
+//                    }
+//                }else {
+//                    poJSON.put("result", "error");
+//                    poJSON.put("message", "No record updated");
+//                    return poJSON;
+//                }
+//            }
+//
+//        } catch (SQLException ex) {
+//            Logger.getLogger(Model_Client_Mobile.class.getName()).log(Level.SEVERE, null, ex);
+//        }
+//        return poJSON;
     }
-
     @Override
     public JSONObject setValue(int lnColumn, Object foValue) {
         
@@ -428,6 +444,5 @@ public class Model_Client_Social_Media implements GEntity{
     public int getEditMode() {
         throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
     }
-    
     
 }
