+++ src/main/java/org/guanzon/cas/model/clients/Model_Client_Master.java
@@ -1,4 +1,4 @@
-package org.guanzon.clients;
+package org.guanzon.cas.model.clients;
 
 import com.google.gson.annotations.SerializedName;
 import java.lang.annotation.Annotation;
