+++ src/main/java/org/guanzon/cas/model/clients/Model_Client_Mobile.java
@@ -1,42 +1,39 @@
-package org.guanzon.clients;
+package org.guanzon.cas.model.clients;
 
 import java.lang.reflect.Method;
 import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import javax.sql.rowset.CachedRowSet;
 import javax.sql.rowset.RowSetProvider;
+import org.guanzon.appdriver.base.CommonUtils;
 import org.guanzon.appdriver.base.GRider;
 import org.guanzon.appdriver.base.MiscUtil;
 import org.guanzon.appdriver.base.SQLUtil;
 import org.guanzon.appdriver.constant.EditMode;
 import org.guanzon.appdriver.constant.Logical;
-import org.guanzon.appdriver.constant.RecordStatus;
 import org.guanzon.appdriver.iface.GEntity;
 import org.json.simple.JSONObject;
-
-
 /**
  *
  * @author Michael Cuison
  */
-public class Model_Client_Mail implements GEntity{
+public class Model_Client_Mobile implements GEntity{
     
-    final String XML = "Model_Client_Mail.xml";
+    final String XML = "Model_Client_Mobile.xml";
     Connection poConn;          //connection
     CachedRowSet poEntity;      //rowset
     String psMessage;           //warning, success or error message
-    
-    
     GRider poGRider;
     int pnEditMode;
-    public JSONObject poJSON;
     
-    public Model_Client_Mail(Connection foValue, GRider poValue){
+    public JSONObject poJSON;
+    public Model_Client_Mobile(Connection foValue, GRider poValue){
         if (foValue == null){
             System.err.println("Database connection is not set.");
             System.exit(1);
@@ -44,12 +41,10 @@ public class Model_Client_Mail implements GEntity{
         pnEditMode = EditMode.UNKNOWN;
         poGRider = poValue;
         poConn = foValue;
-        
         initialize();
     }
 
-    
-        @Override
+    @Override
     public String getColumn(int fnCol) {
         try {
             return poEntity.getMetaData().getColumnLabel(fnCol); 
@@ -70,17 +65,41 @@ public class Model_Client_Mail implements GEntity{
 
     @Override
     public int getColumnCount() {
-        throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
+        try {
+            return poEntity.getMetaData().getColumnCount(); 
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        
+        return -1;
     }
 
-    /** 
-     * this is your table name
-     **/
     @Override
     public String getTable() {
-        return "Client_eMail_Address";
+        return "Client_Mobile";
     }
-     @Override
+
+    @Override
+    public Object getValue(int fnColumn) {
+        try {
+            return poEntity.getObject(fnColumn);
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return null;
+    }
+
+    @Override
+    public Object getValue(String fsColumn) {
+        try {
+            return poEntity.getObject(fsColumn);
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return null;
+    }
+    
+    @Override
     public void list() {
         Method[] methods = this.getClass().getMethods();
         
@@ -90,20 +109,20 @@ public class Model_Client_Mail implements GEntity{
         }
     }
     /**
-     * Sets the Email ID of this record.
+     * Sets the ID of this record.
      * 
      * @param fsValue 
      * @return  True if the record assignment is successful.
      */
-    public JSONObject setEmailID(String fsValue){
-        return setValue("sEmailIDx", fsValue);
+    public JSONObject setMobileID(String fsValue){
+        return setValue("sMobileID", fsValue);
     }
     
-    /**
-     * @return The Email ID of this record. 
+    /** 
+     * @return The ID of this record. 
      */
-    public String getEmailID(){
-        return (String) getValue("sEmailIDx");
+    public String getMobileID(){
+        return (String) getValue("sMobileID");
     }
     
     /**
@@ -124,56 +143,151 @@ public class Model_Client_Mail implements GEntity{
     }
     
     /**
-     * Sets the Entry no.
+     * Sets the contact number.
      * 
      * @param fsValue 
      * @return  True if the record assignment is successful.
      */
-    public JSONObject setEntryNo(int fsValue){
-        return setValue("nEntryNox", fsValue);
+    public JSONObject setContactNo(String fsValue){
+        return setValue("sMobileNo", fsValue);
     }
     
     /**
-     * @return The Entry no.
+     * @return The contact number.
      */
-    public int getEntryNo(){
-        return (int) getValue("nEntryNox");
+    public String getContactNo(){
+        return (String) getValue("sMobileNo");
     }
     
-     /**
-     * Sets the email address.
+    /**
+     * Sets the type of the contact number.<br>
+     * 0 for Mobile No.<br>
+     * 1 for Telephone No.<br>
+     * 2 for Fax Machine No.
      * 
      * @param fsValue 
      * @return  True if the record assignment is successful.
      */
-    public JSONObject setEmail(String fsValue){
-        return setValue("sEMailAdd", fsValue);
+    public JSONObject setContactType(String fsValue){
+        return setValue("cMobileTp", fsValue);
     }
     
     /**
-     * @return The email address.
+     * @return The type of  the contact number. 
      */
-    public String getEmail(){
-        return (String) getValue("sEMailAdd");
+    public String getContactType(){
+        return (String) getValue("cMobileTp");
     }
     
     /**
-     * Sets the priority.
+     * Sets the ownership of the contact number.<br>
+     * 0 for Personal<br>
+     * 1 for Company Issued<br>
+     * 2 for Others
      * 
      * @param fsValue 
      * @return  True if the record assignment is successful.
      */
-    public JSONObject setPriority(int fsValue){
-        return setValue("nPriority", fsValue);
+    public JSONObject setOwnership(String fsValue){
+        return setValue("cOwnerxxx", fsValue);
     }
     
     /**
-     * @return The priority.
+     * @return The ownership of the contact number. 
+     */
+    public String getOwnership(){
+        return (String) getValue("cOwnerxxx");
+    }
+    
+    /**
+     * Sets the contact number as Primary
+     * 
+     * @param fbValue 
+     * @return  True if the record assignment is successful.
+     */
+    public JSONObject setPrimary(boolean fbValue){
+        return setValue("cPrimaryx", fbValue ? "1" : "0");
+    }
+    
+    /**
+     * @return If the contact number set as Primary
+     */
+    public boolean isPrimary(){
+        return ((String) getValue("cPrimaryx")).equals("1");
+    }
+    
+    /**
+     * Sets the priority number of the contact number.
+     * 
+     * @param fnValue 
+     * @return  True if the record assignment is successful.
+     */
+    public JSONObject setPriority(int fnValue){
+        return setValue("nPriority", fnValue);
+    }
+    
+    /**
+     * 
+     * @return The priority number of the contact number. 
      */
     public int getPriority(){
         return (int) getValue("nPriority");
     }
     
+    /**
+     * Sets if the contact number will be receiving marketing message.
+     * 
+     * @param fbValue 
+     * @return  True if the record assignment is successful.
+     */
+    public JSONObject setMarketingRecipient(boolean fbValue){
+        return setValue("cIncdMktg", fbValue ? "1" : "0");
+    }
+    
+    /**
+     * @return If the contact number is receiving marketing message. 
+     */
+    public boolean isMarketingRecipient(){
+        return ((String) getValue("cIncdMktg")).equals("1");
+    }
+    
+    /**
+     * Sets the network if the contact number. Use this method for records as Mobile numbers only.<br>
+     * 0 for Globe Telecom<br>
+     * 1 for Smart Communications<br>
+     * 2 for Sun Cellular<br>
+     * 3 for DITO Telecom
+     * 
+     * @param fsValue 
+     * @return  True if the record assignment is successful.
+     */
+    public JSONObject setMobileNetwork(String fsValue){
+        return setValue("cSubscrbr", fsValue);
+    }
+    
+    /**
+     * @return The network of the contact number. Use this method for records as Mobile numbers only. 
+     */
+    public String getMobileNetwork(){
+        return (String) getValue("cSubscrbr");
+    }
+    
+    /**
+     * Sets record as active.
+     * 
+     * @param fbValue 
+     * @return  True if the record assignment is successful.
+     */
+    public JSONObject setActive(boolean fbValue){
+        return setValue("cRecdStat", fbValue ? "1" : "0");
+    }
+    
+    /**
+     * @return If record is active. 
+     */
+    public boolean isActive(){
+        return ((String) getValue("cRecdStat")).equals("1");
+    }
     
     /**
      * Sets the user encoded/updated the record.
@@ -209,49 +323,64 @@ public class Model_Client_Mail implements GEntity{
         return (Date) getValue("dModified");
     }
     
-    @Override
-    public Object getValue(int fnColumn) {
-        try {
-            return poEntity.getObject(fnColumn);
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-        return null;
+    public String getMessage(){
+        return psMessage;
     }
-
-    @Override
-    public Object getValue(String fsColumn) {
-        try {
-            return poEntity.getObject(fsColumn);
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-        return null;
-    }
-
+    
     private String getSQL(){
         return "SELECT" +
+                    "  sMobileID" +
                     ", sClientID" +
-                    ", nEntryNox" +
-                    ", sEmailAdd" +
+                    ", sMobileNo" +
+                    ", cMobileTp" +
+                    ", cOwnerxxx" +
+                    ", cPrimaryx" +
                     ", nPriority" +
+                    ", cIncdMktg" +
+                    ", nUnreachx" +
+                    ", dLastVeri" +
+                    ", dInactive" +
+                    ", nNoRetryx" +
+                    ", cInvalidx" +
+                    ", cConfirmd" +
+                    ", dConfirmd" +
+                    ", cSubscrbr" +
+                    ", dHoldMktg" +
+                    ", dMktgMsg1" +
+                    ", dMktgMsg2" +
+                    ", dMktgMsg3" +
+                    ", cNewMobil" +
+                    ", cRecdStat" +
+                    ", dModified" +
                 " FROM " + getTable();
     }
+    
     private void initialize(){
         
         try {
             poEntity = MiscUtil.xml2ResultSet(System.getProperty("sys.default.path.metadata") + XML, getTable());
-            
             poEntity.last();
             poEntity.moveToInsertRow();
 
-            MiscUtil.initRowSet(poEntity);  
+            MiscUtil.initRowSet(poEntity);
             
-            
-            poEntity.updateString("sEmailIDx", MiscUtil.getNextCode(getTable(), "sEmailIDx", true, poConn, poGRider.getBranchCode()));
-            poEntity.updateString("cPrimaryx", Logical.NO);
+            //replace with the primary key column info
+//            setClientID(MiscUtil.getNextCode(getTable(), "sMobileID", true, poConn, poGRider.getBranchCode()));
+//            System.out.println("sMobileID = " + getMobileID());
+            poEntity.updateString("sMobileID", MiscUtil.getNextCode(getTable(), "sMobileID", true, poConn, poGRider.getBranchCode()));
+            poEntity.updateString("nUnreachx", Logical.NO);
+            poEntity.updateString("nNoRetryx", Logical.NO);
+            poEntity.updateString("cMobileTp", Logical.NO);
             poEntity.updateString("cOwnerxxx", Logical.NO);
             poEntity.updateString("cRecdStat", Logical.YES);
+            poEntity.updateString("dLastVeri", null);
+            poEntity.updateString("dInactive", null);
+            poEntity.updateString("dConfirmd", null);
+            poEntity.updateString("dHoldMktg", null);
+            poEntity.updateString("dMktgMsg1", null);
+            poEntity.updateString("dMktgMsg2", null);
+            poEntity.updateString("dMktgMsg3", null);
+            
             poEntity.insertRow();
             poEntity.moveToCurrentRow();
 
@@ -260,15 +389,13 @@ public class Model_Client_Mail implements GEntity{
             e.printStackTrace();
             System.exit(1);
         }
-    } 
+    }    
     
-  
+
     @Override
     public JSONObject newRecord() {
         pnEditMode = EditMode.ADDNEW;
         
-        //replace with the primary key column info
-        setEmailID(MiscUtil.getNextCode(getTable(), "sEmailIDx", true, poGRider.getConnection(), poGRider.getBranchCode()));
         
         poJSON = new JSONObject();
         poJSON.put("result", "success");
@@ -276,11 +403,11 @@ public class Model_Client_Mail implements GEntity{
     }
 
     @Override
-    public JSONObject openRecord(String fsValue) { pnEditMode = EditMode.UPDATE;
+    public JSONObject openRecord(String fsValue) {
         poJSON = new JSONObject();
 
         String lsSQL = MiscUtil.makeSelect(this);
-        lsSQL = MiscUtil.addCondition(lsSQL, "sEmailIDx = " + SQLUtil.toSQL(fsValue));
+        lsSQL = MiscUtil.addCondition(lsSQL, "sMobileID = " + SQLUtil.toSQL(fsValue));
 
         ResultSet loRS = poGRider.executeQuery(lsSQL);
 
@@ -314,9 +441,10 @@ public class Model_Client_Mail implements GEntity{
             String lsSQL;
             if (pnEditMode == EditMode.ADDNEW){
                 //replace with the primary key column info
-                setEmailID(MiscUtil.getNextCode(getTable(), "sEmailIDx", true, poGRider.getConnection(), poGRider.getBranchCode()));
-                
+                setMobileID(MiscUtil.getNextCode(getTable(), "sMobileID", true, poGRider.getConnection(), poGRider.getBranchCode()));
                 setModifiedDate(poGRider.getServerDate());
+                
+                setMobileNetwork(CommonUtils.classifyNetwork(getContactNo()));
                 lsSQL = MiscUtil.makeSQL(this);
                 
                 if (!lsSQL.isEmpty()){
@@ -332,15 +460,15 @@ public class Model_Client_Mail implements GEntity{
                     poJSON.put("message", "No record to save.");
                 }
             } else {
-                Model_Client_Mail loOldEntity = new Model_Client_Mail(poConn, poGRider);
+                Model_Client_Mobile loOldEntity = new Model_Client_Mobile(poConn, poGRider);
                 
-                setModifiedDate(poGRider.getServerDate());
                 //replace with the primary key column info
-                JSONObject loJSON = loOldEntity.openRecord(this.getEmailID());
-                
+                JSONObject loJSON = loOldEntity.openRecord(this.getMobileID());
+                setModifiedDate(poGRider.getServerDate());
+                setMobileNetwork(CommonUtils.classifyNetwork(getContactNo()));
                 if ("success".equals((String) loJSON.get("result"))){
                     //replace the condition based on the primary key column of the record
-                    lsSQL = MiscUtil.makeSQL(this, loOldEntity, "sEmailIDx = " + SQLUtil.toSQL(this.getEmailID()));
+                    lsSQL = MiscUtil.makeSQL(this, loOldEntity, "sMobileID = " + SQLUtil.toSQL(this.getClientID()));
                     
                     if (!lsSQL.isEmpty()){
                         if (poGRider.executeQuery(lsSQL, getTable(), poGRider.getBranchCode(), "") > 0){
@@ -364,9 +492,9 @@ public class Model_Client_Mail implements GEntity{
             poJSON.put("message", "Invalid update mode. Unable to save record.");
             return poJSON;
         }
-        
-        return poJSON;
-//
+//        
+//        String lsSQL;
+//        
 //        poJSON =  new JSONObject();
 //        try {
 //            lsSQL = MiscUtil.rowset2SQL(poEntity, 
@@ -375,8 +503,8 @@ public class Model_Client_Mail implements GEntity{
 //                    "");
 //        
 //            if (pnEditMode == EditMode.ADDNEW){           
-//                lsSQL = MiscUtil.getNextCode(getTable(), "sEmailIDx", false, poGRider.getConnection(), "");
-//                poEntity.updateObject("sEmailIDx", lsSQL);
+//                lsSQL = MiscUtil.getNextCode(getTable(), "sMobileID", false, poGRider.getConnection(), "");
+//                poEntity.updateObject("sMobileID", lsSQL);
 //                poEntity.updateRow();
 //
 //                lsSQL = MiscUtil.rowset2SQL(poEntity, getTable(), "");
@@ -384,7 +512,7 @@ public class Model_Client_Mail implements GEntity{
 //                lsSQL = MiscUtil.rowset2SQL(poEntity, 
 //                                            getTable(), 
 //                                            "", 
-//                                            "sEmailIDx = " + SQLUtil.toSQL(poEntity.getString("sEmailIDx")));
+//                                            "sMobileID = " + SQLUtil.toSQL(poEntity.getString("sMobileID")));
 //            }
 //            
 //            if (!lsSQL.equals("")){
@@ -404,8 +532,10 @@ public class Model_Client_Mail implements GEntity{
 //        } catch (SQLException ex) {
 //            Logger.getLogger(Model_Client_Mobile.class.getName()).log(Level.SEVERE, null, ex);
 //        }
-//        return poJSON;
+        return poJSON;
+
     }
+
     @Override
     public JSONObject setValue(int lnColumn, Object foValue) {
         
@@ -444,5 +574,6 @@ public class Model_Client_Mail implements GEntity{
     public int getEditMode() {
         throw new UnsupportedOperationException("Not supported yet."); // Generated from nbfs://nbhost/SystemFileSystem/Templates/Classes/Code/GeneratedMethodBody
     }
+    
     
 }
